+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - fix: drag a node without comboId by drag-node with onlyChangeComboSize;
 - fix: gpu layout with async;
 - fix: minimap with delegate type cannot reach the top of the canvas, closes: #2885;
+- feat: improve the performance for updating nodes;
+- feat: updateLayout with align and alignPoint;
 
 #### 4.3.4